# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample API for DigitialReserve portlet",
    "version": "1.0.0",
    "title": "Digital Reserve",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "hgarc@bu.edu"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "tags": [
    {
      "name": "subject",
      "description": "All subject operations",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://www.hgar-srv3.bu.edu"
      }
    },
    {
      "name": "user",
      "description": "all user operations",
      "externalDocs": {
        "description": "Find out more about the user",
        "url": "http://www.hgar-srv3.bu.edu"
      }
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/subject": {
      "post": {
        "tags": [
          "subject"
        ],
        "summary": "Add a new subject to the portlet",
        "description": "adds a subject to the list in the portlet",
        "operationId": "addSubject",
        "parameters": [
          {
            "name": "subjectURL",
            "in": "query",
            "type": "string",
            "description": "Subject/Item object that will be added in the store",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Something went wrong"
          }
        }
      },
      "delete": {
        "tags": [
          "subject"
        ],
        "summary": "Deletes a subject",
        "description": "",
        "operationId": "deleteSubject",
        "parameters": [
          {
            "name": "subjectID",
            "in": "query",
            "description": "Subject id to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Subject not found"
          }
        }
      }
    },
    "/subject/{itemID}": {
      "post": {
        "tags": [
          "subject"
        ],
        "summary": "adds an item to a subject",
        "description": "gets an ID and adds it to the subject",
        "operationId": "addItemInSubject",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "itemID",
            "in": "path",
            "description": "item's identification",
            "required": true,
            "type": "string"
          },
          {
            "name": "subjectID",
            "in": "query",
            "description": "subject's ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Subject not found"
          }
        }
      },
      "delete": {
        "tags": [
          "subject"
        ],
        "summary": "deletes an item in a subject",
        "description": "delete an item using its identification",
        "operationId": "deleteItemInSubject",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "itemID",
            "in": "path",
            "description": "item's identification",
            "required": true,
            "type": "string"
          },
          {
            "name": "subjectID",
            "in": "query",
            "description": "subject's ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Subject not found"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "display all the saved subjects for a logged in used",
        "description": "the user will be able to get all the saved material in display in the portlet",
        "operationId": "displayContent",
        "parameters": [
          {
            "name": "userID",
            "in": "query",
            "type": "string",
            "required": true,
            "description": "the user's identification in the database"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User doesn't exist"
          }
        }
      }
    }
  }
}